app/requirements.txt
beautifulsoup4
requests
redis
APScheduler
selenium
praw
twitter
facebook-sdk